separate caveat file into respective caveat files
